Extract writing logic to new output module in application crate
